refactor(be): disable absolute path import
