+++ backend/apps/client/src/auth/auth.module.ts
@@ -2,8 +2,7 @@ import { Module } from '@nestjs/common'
 import { ConfigModule, ConfigService } from '@nestjs/config'
 import { JwtModule, type JwtModuleOptions } from '@nestjs/jwt'
 import { PassportModule } from '@nestjs/passport'
-import { GithubStrategy } from 'libs/auth/src/github/github.strategy'
-import { JwtAuthModule } from '@libs/auth'
+import { JwtAuthModule, GithubStrategy } from '@libs/auth'
 import { UserModule } from '@client/user/user.module'
 import { AuthController } from './auth.controller'
 import { AuthService } from './auth.service'
