+++ backend/libs/constants/src/index.ts
@@ -1,3 +1,4 @@
+export * from './oauth.constants'
 export * from './time.constants'
 export * from './rabbitmq.constants'
 export * from './submission.constants'
