+++ backend/libs/pipe/src/index.ts
@@ -1,2 +1,4 @@
 export * from './cursor-validation.pipe'
 export * from './zod-validation.pipe'
+export * from './id-validation.pipe'
+export * from './group-id.pipe'
