+++ backend/tsconfig.json
@@ -9,7 +9,6 @@
     "target": "esnext",
     "sourceMap": true,
     "outDir": "./dist",
-    "baseUrl": ".",
     "incremental": true,
     "skipLibCheck": true,
     "strictNullChecks": true,
@@ -18,17 +17,17 @@
     "forceConsistentCasingInFileNames": false,
     "noFallthroughCasesInSwitch": false,
     "paths": {
-      "@admin/*": ["apps/admin/src/*"],
-      "@client/*": ["apps/client/src/*"],
-      "@generated": ["apps/admin/src/@generated"],
-      "@libs/prisma": ["libs/prisma/src/index.ts"],
-      "@libs/cache": ["libs/cache/src/index.ts"],
-      "@libs/auth": ["libs/auth/src/index.ts"],
-      "@libs/exception": ["libs/exception/src/index.ts"],
-      "@libs/pipe": ["libs/pipe/src/index.ts"],
-      "@libs/constants": ["libs/constants/src/index.ts"],
-      "@libs/logger": ["libs/logger/src/index.ts"],
-      "@libs/types": ["libs/types/src/index.ts"]
+      "@admin/*": ["./apps/admin/src/*"],
+      "@client/*": ["./apps/client/src/*"],
+      "@generated": ["./apps/admin/src/@generated"],
+      "@libs/prisma": ["./libs/prisma/src/index.ts"],
+      "@libs/cache": ["./libs/cache/src/index.ts"],
+      "@libs/auth": ["./libs/auth/src/index.ts"],
+      "@libs/exception": ["./libs/exception/src/index.ts"],
+      "@libs/pipe": ["./libs/pipe/src/index.ts"],
+      "@libs/constants": ["./libs/constants/src/index.ts"],
+      "@libs/logger": ["./libs/logger/src/index.ts"],
+      "@libs/types": ["./libs/types/src/index.ts"]
     }
   }
 }
